Give up on StopIteration exception
Since wait times are defined by generators, it makes sense that if
a generator runs out of values that would imply giving up.  Since
all the built in generators are infinite this hasn't come up before,
but it should be possible to define a custom generator that yields a
finite number of values and implicitly triggers a giveup once it is
exhausted.