Support custom loggers
This adds a `logger` kwarg to both decorators.  This allows the user
to specify an alternative logger to use for the backoff and giveup
events. This logger can be specified either as a string name or as a
Logger object.

In addition, the default logging can be disabled entirely by
specifying logger=None.  In this case, a user could specify their own
alterative logging behavior by way of event handlers.

This should address multiple issues that have been filed around making
logging more customizable, including:

https://github.com/litl/backoff/issues/37
https://github.com/litl/backoff/issues/50
https://github.com/litl/backoff/issues/52