Add protection against max_time bug
When I specify the `max_time` argument, I expect the worst-case upper
bound for execution time to be the specified timeout + the execution
time of a single function call. Moreover, I would not expect retries to
trigger if the give up condition is already fulfilled after the first
function execution.

The existing behavior appears to be caused by collected `elapsed` for
the `details` before the target function is executed. Both of the above
bugs should be fixed by swapping that ordering.