Handle wait_gen_kwargs which take an argument
All wait_gen args are subject to maybe_call() as a way to dynamically
look up a value at runtime.  These callables do not take an argument.
If a callable which takes an argument is passed (as is the case with
the `value` kwargs for the `runtime` wait generator, we pass through
the function.