Refrain from adding Details to backoff namespace
If needed, I want to make sure we put this in the right place. There
may ultimately be some other types which it would be helpful to expose
as well, and in that case I think it might be better to have a typing
submodule where only these types are exposed.